Reorganize lark logic
